+++ src/components/search/ResultItem.vue
@@ -25,41 +25,11 @@
 					</div>
 					<div class="result-image-wrapper"><kb-imagecomponent :imagedata="imageData"></kb-imagecomponent></div>
 				</div>
-				<div class="extra-features">
-					<button
-						:disabled="resultdata && resultdata?.file_id?.length > 0 ? false : true"
-						class="thumbnail-button"
-						:title="'See more thumbnails'"
-						@click="showThumbnails()"
-					>
-						<span class="material-icons">photo_library</span>
-					</button>
-				</div>
-				<div
-					ref="extraContentRef"
-					class="extra-content"
-				>
-					<ItemSlider item-class="extra-thumbnail">
-						<template #default="slotProps">
-							<router-link
-								v-for="(item, index) in thumbnailImages"
-								:key="index"
-								:to="{ path: 'record/' + resultdata.id }"
-								role="link"
-								class="extra-thumbnail"
-								v-bind="slotProps"
-							>
-								<div
-									ref="thumbnailRefs"
-									class="img-wrap"
-								>
-									<kb-imagecomponent :imagedata="thumbnailImageData[index]"></kb-imagecomponent>
-								</div>
-								<div class="img-stamp">timestamp</div>
-							</router-link>
-						</template>
-					</ItemSlider>
-				</div>
+				<AdditionalInfo
+					:id="resultdata.id"
+					:file-id="resultdata.file_id ? resultdata.file_id : ''"
+					:duration="Number(resultdata.duration_ms)"
+				></AdditionalInfo>
 			</div>
 			<div
 				v-else
@@ -104,15 +74,14 @@ import { useSearchResultStore } from '@/store/searchResultStore';
 import { GenericSearchResultType } from '@/types/GenericSearchResultTypes';
 import { ImageComponentType } from '@/types/ImageComponentType';
 import { APIService } from '@/api/api-service';
-import ItemSlider from '@/components/search/ItemSlider.vue';
-import gsap from 'gsap';
+import AdditionalInfo from '@/components/search/AdditionalInfo.vue';
 
 import '@/components/common/wc-image-item';
 
 export default defineComponent({
 	name: 'ResultItem',
 	components: {
-		ItemSlider,
+		AdditionalInfo,
 	},
 	props: {
 		resultdata: {
@@ -142,10 +111,6 @@ export default defineComponent({
 	},
 	setup(props) {
 		const searchResultStore = useSearchResultStore();
-		const thumbnailImages = ref(10);
-		const extraContentShown = ref(false);
-		const thumbnailImageData = ref([] as string[]);
-
 		//Default imageData obj to prevent render issues
 		const imageData = ref(
 			JSON.stringify({
@@ -158,8 +123,6 @@ export default defineComponent({
 		const placeholderSummaryRefs = ref<HTMLElement | []>([]);
 		const placeholderSubtitleRefs = ref<HTMLElement | []>([]);
 		const placeholderTitleRef = ref<HTMLElement | null>(null);
-		const extraContentRef = ref<HTMLElement | null>(null);
-		const thumbnailRefs = ref<HTMLAnchorElement[]>([]);
 
 		const getImageData = () => {
 			const imageDataObj = {} as ImageComponentType;
@@ -195,64 +158,9 @@ export default defineComponent({
 			(newVal, oldVal) => {
 				if (newVal !== oldVal) {
 					getImageData();
-					thumbnailImageData.value = [];
-					extraContentShown.value = false;
 				}
 			},
 		);
-
-		const showThumbnails = () => {
-			extraContentShown.value = !extraContentShown.value;
-			if (props.resultdata?.file_id && extraContentShown.value) {
-				if (thumbnailImageData.value.length === 0) {
-					requestExtraThumbnails();
-				}
-			}
-			if (extraContentShown.value === true) {
-				gsap.set(extraContentRef.value, {
-					visibility: 'visible',
-				});
-			}
-			gsap.to(extraContentRef.value, {
-				height: extraContentShown.value ? 'auto' : '0px',
-				opacity: extraContentShown.value ? '1' : '0',
-				duration: 0.2,
-				onComplete: () => {
-					if (extraContentShown.value === false) {
-						gsap.set(extraContentRef.value, {
-							visibility: 'hidden',
-						});
-					}
-				},
-			});
-		};
-
-		const requestExtraThumbnails = () => {
-			APIService.getExtraThumbnails(props.resultdata.file_id)
-				.then((thumbServiceResponse) => {
-					console.log(thumbServiceResponse);
-
-					const img = new Image();
-					img.src = thumbServiceResponse.data.sprite;
-					img.onload = () => {
-						thumbnailRefs.value.forEach((item) => {
-							item.style.height = img.height + 'px';
-						});
-					};
-
-					thumbnailRefs.value.forEach((item, index) => {
-						const imgData = {} as ImageComponentType;
-						imgData.imgSrc = thumbServiceResponse.data.sprite;
-						imgData.objectPos = `-${200 * index}px 0px`;
-						imgData.imgOption = 'none';
-						thumbnailImageData.value.push(JSON.stringify(imgData));
-					});
-				})
-				//Just in case the service fail - we fail silently and swoop in with the placeholder
-				.catch(() => {
-					//nay
-				});
-		};
 
 		onMounted(() => {
 			getImageData();
@@ -264,12 +172,6 @@ export default defineComponent({
 			placeholderSubtitleRefs,
 			placeholderSummaryRefs,
 			placeholderTitleRef,
-			extraContentRef,
-			showThumbnails,
-			extraContentShown,
-			thumbnailImages,
-			thumbnailRefs,
-			thumbnailImageData,
 			toRaw,
 		};
 	},
@@ -469,66 +371,6 @@ export default defineComponent({
 	max-height: calc(20px * 3); /* fallback for firefox */
 }
 
-.extra-features {
-	margin-top: 10px;
-}
-
-.extra-features .material-icons {
-	font-size: 20px;
-}
-
-.thumbnail-button {
-	cursor: pointer;
-	border: 1px solid #002f702a;
-	border-radius: 0px;
-	background-color: transparent;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	color: #002e70;
-	transition: all 0.3s ease-in-out 0s;
-}
-
-.thumbnail-button:hover {
-	border: 1px solid #002e70;
-}
-
-.thumbnail-button:disabled:hover {
-	border: 1px solid #002f702a;
-	cursor: auto;
-}
-
-.thumbnail-button:disabled {
-	color: rgba(180, 180, 180, 1);
-}
-
-.extra-content {
-	height: 0px;
-}
-
-.extra-thumbnail {
-	flex: 0 0 200px;
-	position: relative;
-	z-index: 1;
-	object-fit: none;
-	display: flex;
-	flex-direction: column;
-	pointer-events: all;
-}
-
-.extra-thumbnail.disabled {
-	pointer-events: none;
-}
-
-.extra-thumbnail .img-wrap {
-	margin-bottom: 10px;
-	height: 105px;
-}
-
-.extra-thumbnail .img-stamp {
-	text-align: center;
-	font-size: 12px;
-}
 @media (min-width: 800px) {
 	.title {
 		max-width: calc(100% - (200px - 60px));
