+++ src/types/GenericSearchResultTypes.ts
@@ -33,7 +33,7 @@ export interface GenericSearchResultType {
 	_version: number;
 	startTime?: string;
 	endTime?: string;
-	duration?: string;
+	duration_ms: string;
 	origin: string;
 }
