+++ src/utils/time-utils.ts
@@ -70,6 +70,21 @@ function generateDurationParts(hours: number, minutes: number, seconds: number, 
 	return durationParts;
 }
 
+function getTimeFromStartAndEnde(duration: string) {
+	const match = duration.match(/PT(\d+H)?(\d+M)?(\d+S)?/);
+	if (match) {
+		const hours = parseInt(match[1] || '0');
+		const minutes = parseInt(match[2] || '0');
+		const seconds = parseInt(match[3] || '0');
+
+		const totalMilliseconds = ((hours * 60 + minutes) * 60 + seconds) * 1000;
+
+		return totalMilliseconds;
+	} else {
+		return 0;
+	}
+}
+
 function getBroadcastDate(isoDate: string, locale: string): string {
 	const date = new Date(isoDate);
 
@@ -97,4 +112,4 @@ function getBroadcastTime(isoDate: string): string {
 	return new Intl.DateTimeFormat('en-GB', options).format(dateObj);
 }
 
-export { formatDuration, getBroadcastDate, getBroadcastTime };
+export { formatDuration, getBroadcastDate, getBroadcastTime, getTimeFromStartAndEnde };
